fix(overlap): extract ink helpers + count Text subclasses correctly (0.29.17)
Two bundled changes on _quality/_overlap.py (residual from the 0.29.11
annotation-overlap fix, which sat at the 512-line ceiling):

1. EXTRACT the data-only "ink" raster helpers (_render_ink_mask,
   _ink_fraction_in_bbox, _parse_bg_color, _BG_TOL) into a new
   _quality/_overlap_ink.py. _overlap.py re-imports the two used by the
   detector and drops from 512 -> 422 lines (thin orchestrator).
2. FIX the ink-mask text-hide: _render_ink_mask hid artists by exact class
   name "Text", so Annotation (a Text SUBCLASS) leaked into the data-only ink
   mask — annotations were wrongly counted as data ink for the legend-vs-ink
   check. Now isinstance(matplotlib.text.Text). Also converted the enumeration
   MRO-check to the idiomatic isinstance for consistency.

Closes card figrecipe-overlap-inkmask-extraction. Per operator 2026-07-08
(autonomous go-ahead / constitution).

Tests: annotation-not-counted-as-data-ink (new); existing 10 overlap tests +
scatter_labels (which re-imports _render_ink_mask) stay green.
